removed redundant ObservationCategoryMap interface
