+++ api/src/main/java/org/openmrs/module/fhir2/api/mappings/BaseMapping.java
@@ -7,7 +7,7 @@
  * Copyright (C) OpenMRS Inc. OpenMRS is a registered trademark and the OpenMRS
  * graphic logo is a trademark of OpenMRS Inc.
  */
-package org.openmrs.module.fhir2.api.mappings.impl;
+package org.openmrs.module.fhir2.api.mappings;
 
 import java.io.IOException;
 import java.io.InputStream;
